cleanup example script
