+++ apps/sdk-sandbox/scripts/example.ts
@@ -2,11 +2,11 @@ import { Drift } from "@delvtech/drift";
 import { viemAdapter } from "@delvtech/drift-viem";
 import { appConfig } from "@delvtech/hyperdrive-appconfig";
 import { ReadHyperdrive, ReadWriteHyperdrive } from "@delvtech/hyperdrive-js";
-import { CurrencyAmount, Percent, Token, TradeType } from "@uniswap/sdk-core";
-import { AlphaRouter, SwapType } from "@uniswap/smart-order-router";
-import { ethers } from "ethers";
+import { Token } from "@uniswap/sdk-core";
 import { publicClient, walletClient } from "../client";
 import { executeZapOpenAndClose } from "./executeZapAndClose";
+import { fetchSwapPath } from "./fetchSwapPath";
+
 const zapsConfig = appConfig.zaps[707];
 const drift = new Drift(viemAdapter({ publicClient, walletClient }));
 
@@ -28,75 +28,10 @@ const readPool = new ReadHyperdrive({
   earliestBlock,
 });
 
-const poolContract = drift.contract({
+drift.contract({
   abi: writePool.contract.abi,
   address: poolAddress,
 });
-
-async function fetchSwapPath(tokenIn: Token, tokenOut: Token) {
-  const provider = new ethers.providers.JsonRpcProvider(process.env.RPC_URL);
-  const router = new AlphaRouter({
-    chainId: 1,
-    provider,
-  });
-
-  // Swap amount: 100 DAI
-  const amountIn = ethers.utils.parseUnits("100", 18);
-  const currencyAmountIn = CurrencyAmount.fromRawAmount(
-    tokenIn,
-    amountIn.toString(),
-  );
-
-  // Fetch the route for an exact input swap
-  const route = await router.route(
-    currencyAmountIn,
-    tokenOut,
-    TradeType.EXACT_INPUT,
-    {
-      recipient: "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266", // replace with your address
-      slippageTolerance: new Percent(50, 10000), // 0.50% tolerance
-      deadline: Math.floor(Date.now() / 1000) + 1800, // 30 min deadline
-      type: SwapType.SWAP_ROUTER_02,
-    },
-  );
-
-  if (route) {
-    // Extract the V3 route (which contains tokenPath and poolIdentifiers)
-    const v3Route = route.route.find((r) => r.protocol === "V3");
-
-    if (v3Route) {
-      const tokenPath = v3Route.tokenPath;
-      const pools = v3Route.route.pools; // each pool contains its fee
-      console.log(
-        "V3 tokenPath:",
-        tokenPath.map((t) => t.address),
-      );
-
-      // Build arrays of types and values for solidityPack
-      const types: string[] = [];
-      const values: (string | number)[] = [];
-
-      for (let i = 0; i < tokenPath.length; i++) {
-        // Add token address
-        types.push("address");
-        values.push(tokenPath[i].address);
-        // For every hop except the last, add the fee from the corresponding pool
-        if (i < tokenPath.length - 1) {
-          const fee = pools[i].fee;
-          types.push("uint24");
-          values.push(fee);
-        }
-      }
-      const encodedPath = ethers.utils.solidityPack(types, values);
-      console.log("EncodedPath", encodedPath);
-      return encodedPath;
-    } else {
-      console.log("No V3 route found in the swap path.");
-    }
-  } else {
-    console.log("No swap path found.");
-  }
-}
 
 async function main() {
   const DAI = new Token(
@@ -113,7 +48,7 @@ async function main() {
     "USDC",
     "USD Coin",
   );
-  // await executeZapOpenAndClose();
+
   const encodedPath = await fetchSwapPath(DAI, USDC);
   if (!encodedPath) {
     console.log("No encoded path found");
