+++ apps/sdk-sandbox/scripts/executeZapAndClose.ts
@@ -16,8 +16,6 @@ const drift = new Drift(viemAdapter({ publicClient, walletClient }));
 const poolAddress = "0x324395D5d835F84a02A75Aa26814f6fD22F25698";
 const earliestBlock = 20180617n;
 const usdcAddress = "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48";
-const daiAddress = "0x6b175474e89094c44da98b954eedeac495271d0f";
-// SAMPLE ASSET ID AND MATURITY
 const assetId: bigint =
   452312848583266388373324160190187140051835877600158453279131187532667606656n;
 const maturity = 1756944000n;
@@ -31,7 +29,6 @@ const writePool = new ReadWriteHyperdrive({
   earliestBlock,
 });
 
-// Read instance (includes zapAddress)
 const readPool = new ReadHyperdrive({
   address: poolAddress,
   drift,
@@ -50,14 +47,7 @@ export async function executeZapOpenAndClose(swapPath: `0x${string}`) {
     const account = walletClient?.account.address as Address;
     console.log("[START] Account:", account);
 
-    // Get initial open long details
-    const beforeDetails = await readPool.getOpenLongDetails({
-      account,
-      assetId,
-      options: { block: "latest" },
-    });
-    // console.log("openLongDetails BEFORE openLong:", beforeDetails);
-
+    // Log initial balances
     const beforeBalanceOfBonds = await readPool.contract.read("balanceOf", {
       owner: account,
       tokenId: assetId,
@@ -71,112 +61,23 @@ export async function executeZapOpenAndClose(swapPath: `0x${string}`) {
       args: [account],
     });
     console.log("beforeBalanceOfUsdc", fixed(beforeBalanceOfUsdc, 6).format());
-    // Open Long
-    // const { result, request } = await publicClient.simulateContract({
-    //   abi: writePool.contract.abi,
-    //   address: poolAddress,
-    //   chain: publicClient.chain,
-    //   functionName: "openLong",
-    //   account,
-    //   gas: 16125042n,
-    //   args: [
-    //     defaultOpenLongAmount, // 20 base token
-    //     1n,
-    //     1n,
-    //     {
-    //       asBase: true,
-    //       destination: account,
-    //       extraData: "0x",
-    //     },
-    //   ],
-    // });
 
-    // const openTxHash = await walletClient?.writeContract({
-    //   ...request,
-    // });
-    // if (!openTxHash) throw new Error("No open transaction hash received");
-    // console.log("Open Long tx hash:", openTxHash);
-
-    // const txReceipt = await publicClient.waitForTransactionReceipt({
-    //   hash: openTxHash,
-    // });
-    // await drift.cache.clear();
-
-    // if (!txReceipt) throw new Error("No open transaction receipt received");
-    // if (txReceipt.status !== "success") {
-    //   console.error("Open Transaction failed:", txReceipt);
-    //   throw new Error("Open Transaction failed");
-    // }
-    // console.log("Open tx receipt status:", txReceipt.status);
-
-    // // Get open long details after opening
-    // const afterDetails = await readPool.getOpenLongDetails({
-    //   account,
-    //   assetId,
-    // });
-
-    // await drift.cache.clear();
-    // const afterBalanceOfBonds = await readPool.contract.read("balanceOf", {
-    //   owner: account,
-    //   tokenId: assetId,
-    // });
-    // console.log("afterBalanceOfBonds", fixed(afterBalanceOfBonds).format());
-
-    // Check the USDC Balance here
-    const usdcBalanceOf = await publicClient.readContract({
-      address: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
-      abi: erc20Abi,
-      functionName: "balanceOf",
-      args: [account],
-    });
-    console.log("usdcBalance", fixed(usdcBalanceOf, 6).format());
-
-    // console.log("openLongDetails AFTER openLong:", afterDetails);
-
-    // Approve zap (auxiliary) contract to manage the long position
-    // const approvalReceipt = await writePool.contract.write(
-    //   "setApprovalForAll",
-    //   {
-    //     operator: zapsConfig.address,
-    //     approved: true,
-    //   },
-    // );
-
+    // Approve zap contract to manage the long position
     const approveAssetId = await writePool.contract.write("setApproval", {
       operator: zapsConfig.address,
       tokenID: assetId,
       amount: maxInt256,
     });
-    // await publicClient.waitForTransactionReceipt({
-    //   hash: approvalReceipt,
-    // });
-    await publicClient.waitForTransactionReceipt({
-      hash: approveAssetId,
+    await publicClient.waitForTransactionReceipt({ hash: approveAssetId });
+
+    const revokeApproval = await writePool.contract.write("setApprovalForAll", {
+      operator: zapsConfig.address,
+      approved: false,
     });
+    await publicClient.waitForTransactionReceipt({ hash: revokeApproval });
+    console.log("Approval tx hash:", revokeApproval);
 
-    // const allowanceOfZapsToSpend = await publicClient.readContract({
-    //   address: poolAddress,
-    //   abi: poolContract.abi,
-    //   functionName: "isApprovedForAll",
-    //   args: [account, zapsConfig.address],
-    // });
-    // console.log("allowanceOfZapsToSpend", allowanceOfZapsToSpend);
-
-    const approvalReceipt = await writePool.contract.write(
-      "setApprovalForAll",
-      {
-        operator: zapsConfig.address,
-        approved: false,
-      },
-    );
-
-    await publicClient.waitForTransactionReceipt({
-      hash: approvalReceipt,
-    });
-
-    console.log("Approval tx hash:", approvalReceipt);
-
-    // Simulate closeLong to preview base amount out
+    // Preview closeLong base amount out
     const { result: previewBaseAmountOut } =
       await publicClient.simulateContract({
         abi: writePool.contract.abi,
@@ -186,7 +87,7 @@ export async function executeZapOpenAndClose(swapPath: `0x${string}`) {
         account,
         args: [
           maturity,
-          defaultCloseLongAmount, // 10 base token (example)
+          defaultCloseLongAmount,
           1n,
           {
             asBase: true,
@@ -195,83 +96,39 @@ export async function executeZapOpenAndClose(swapPath: `0x${string}`) {
           },
         ],
       });
-    // console.log("Preview base out (raw):", previewBaseAmountOut);
-    console.log(
-      "Preview base out (formatted):",
-      fixed(previewBaseAmountOut).format(),
-    );
+    console.log("Preview base out:", fixed(previewBaseAmountOut).format());
 
-    // Prepare zap swap parameters
-    const blockData = await publicClient.getBlock();
-    const deadline = blockData.timestamp + 60n;
+    // Prepare swap parameters
+    const { timestamp } = await publicClient.getBlock();
+    const deadline = timestamp + 60n;
 
-    // const simulateReadNameFromZap = await publicClient.readContract({
-    //   address: zapsConfig.address,
-    //   abi: zapAbi,
-    //   functionName: "name",
-    // });
-    // console.log("simulateReadNameFromZap", simulateReadNameFromZap);
-
-    // const simulateCloseLongZap = await publicClient.simulateContract({
-    //   abi: zapAbi,
-    //   address: zapsConfig.address,
-    //   functionName: "closeLongZap",
-    //   gas: 20000000n, // Increased gas limit
-    //   account,
-    //   args: [
-    //     poolAddress,
-    //     maturity,
-    //     defaultCloseLongAmount,
-    //     1n,
-    //     {
-    //       destination: zapsConfig.address,
-    //       asBase: true,
-    //       extraData: "0x",
-    //     },
-    //     {
-    //       path: swapPath,
-    //       recipient: account,
-    //       deadline,
-    //       amountIn: previewBaseAmountOut,
-    //       amountOutMinimum: BigInt(1e6),
-    //     },
-    //     false,
-    //   ],
-    // });
-    // console.log("simulateCloseLongZap", simulateCloseLongZap);
-    const swapTx = await walletClient
-      ?.writeContract({
-        abi: zapAbi,
-        chain: publicClient.chain,
-        address: zapsConfig.address,
-        functionName: "closeLongZap",
-        gas: 20000000n, // Increased gas limit
-        args: [
-          poolAddress,
-          maturity,
-          defaultCloseLongAmount, // 10 base token
-          0n,
-          {
-            // hyperdriver parameters
-            destination: zapsConfig.address,
-            asBase: true,
-            extraData: "0x",
-          },
-          {
-            // Swap parameters
-            amountIn: previewBaseAmountOut,
-            amountOutMinimum: 1n,
-            deadline,
-            path: swapPath,
-            recipient: account,
-          },
-          false,
-        ],
-      })
-      .catch((err) => {
-        console.error("closeLongZap failed:", err);
-        throw err;
-      });
+    // Execute zap swap
+    const swapTx = await walletClient?.writeContract({
+      abi: zapAbi,
+      chain: publicClient.chain,
+      address: zapsConfig.address,
+      functionName: "closeLongZap",
+      gas: 20000000n,
+      args: [
+        poolAddress,
+        maturity,
+        defaultCloseLongAmount,
+        0n,
+        {
+          destination: zapsConfig.address,
+          asBase: true,
+          extraData: "0x",
+        },
+        {
+          amountIn: previewBaseAmountOut,
+          amountOutMinimum: 1n,
+          deadline,
+          path: swapPath,
+          recipient: account,
+        },
+        false,
+      ],
+    });
     if (!swapTx) throw new Error("No close position transaction hash received");
     console.log("closeLongZap tx hash:", swapTx);
 
@@ -279,15 +136,10 @@ export async function executeZapOpenAndClose(swapPath: `0x${string}`) {
       hash: swapTx,
     });
     console.log("Receipt status after zap close:", receipt.status);
-    await drift.cache.clear();
-
-    // const openLongDetailsAfterZap = await readPool.getOpenLongDetails({
-    //   account,
-    //   assetId,
-    // });
-    // console.log("openLongDetails AFTER zap close:", openLongDetailsAfterZap);
 
     await drift.cache.clear();
+
+    // Log final balances
     const afterZapBalanceOfBonds = await readPool.contract.read("balanceOf", {
       owner: account,
       tokenId: assetId,
@@ -298,14 +150,12 @@ export async function executeZapOpenAndClose(swapPath: `0x${string}`) {
     );
 
     const usdcBalanceAfterZap = await publicClient.readContract({
-      address: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
+      address: usdcAddress,
       abi: erc20Abi,
       functionName: "balanceOf",
       args: [account],
     });
     console.log("usdcBalanceAfterZap", fixed(usdcBalanceAfterZap, 6).format());
-
-    // return receipt;
   } catch (error) {
     console.error("Failed to execute zap open/close:", error);
     throw error;
